+++ .changeset/wicked-dolls-pull.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/aleno-adapter': minor
+---
+
+Remove REST transport. Will fall back to SocketIO transport even if REST is requested.
