+++ packages/sources/aleno/src/endpoint/price.ts
@@ -1,11 +1,9 @@
 import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
 import { InputParameters } from '@chainlink/external-adapter-framework/validation'
 import { SingleNumberResultResponse } from '@chainlink/external-adapter-framework/util'
-import { TransportRoutes } from '@chainlink/external-adapter-framework/transports'
 import { priceEndpointInputParametersDefinition } from '@chainlink/external-adapter-framework/adapter'
 import { config } from '../config'
 
-import { httpTransport } from '../transport/price-http'
 import { socketioTransport } from '../transport/price-socketio'
 
 export const inputParameters = new InputParameters(priceEndpointInputParametersDefinition, [
@@ -24,9 +22,7 @@ export type BaseEndpointTypes = {
 export const endpoint = new AdapterEndpoint({
   name: 'price',
   aliases: ['crypto', 'state'],
-  transportRoutes: new TransportRoutes<BaseEndpointTypes>()
-    .register('rest', httpTransport)
-    .register('ws', socketioTransport),
+  transport: socketioTransport,
   inputParameters,
   defaultTransport: 'ws',
 })
