--- packages/sources/aleno/src/transport/price-http.ts
@@ -1,94 +1,0 @@
-import { HttpTransport } from '@chainlink/external-adapter-framework/transports'
-import { BaseEndpointTypes } from '../endpoint/price'
-
-interface ResponseItem {
-  id: string
-  baseSymbol: string
-  quoteSymbol: string
-  processTimestamp: number
-  processBlockChainId: string
-  processBlockNumber: number
-  processBlockTimestamp: number
-  aggregatedLast7DaysBaseVolume: number
-  price: number
-  aggregatedMarketDepthMinusOnePercentUsdAmount: number
-  aggregatedMarketDepthPlusOnePercentUsdAmount: number
-  aggregatedMarketDepthUsdAmount: number
-  aggregatedLast7DaysUsdVolume: number
-}
-
-export type ResponseSchema = ResponseItem[]
-
-export type HttpTransportTypes = BaseEndpointTypes & {
-  Provider: {
-    RequestBody: never
-    ResponseBody: ResponseSchema
-  }
-}
-
-export const httpTransport = new HttpTransport<HttpTransportTypes>({
-  prepareRequests: (params, config) => {
-    return params.map((param) => {
-      return {
-        params: [param],
-        request: {
-          baseURL: config.API_ENDPOINT,
-          url: 'baseTokenStates/latest',
-          headers: {
-            'x-api-key': config.API_KEY,
-          },
-        },
-      }
-    })
-  },
-
-  parseResponse: (params, response) => {
-    if (!response.data) {
-      return params.map((param) => {
-        return {
-          params: param,
-          response: {
-            errorMessage: `The data provider didn't return any value for ${param.base}/${param.quote}`,
-            statusCode: 502,
-          },
-        }
-      })
-    }
-
-    return params.map((param) => {
-      let result
-      let processTimestamp
-
-      response.data.forEach((row) => {
-        if (row.baseSymbol === param.base && row.quoteSymbol === param.quote) {
-          result = Number(row.price)
-          processTimestamp = row.processTimestamp
-        }
-      })
-
-      if (result === undefined || processTimestamp === undefined) {
-        return {
-          params: param,
-          response: {
-            errorMessage: `The data provider didn't return any value for ${param.base}/${param.quote}`,
-            statusCode: 502,
-          },
-        }
-      }
-
-      return {
-        params: param,
-        response: {
-          result,
-          data: {
-            result,
-          },
-          timestamps: {
-            providerDataReceivedUnixMs: Date.now(),
-            providerIndicatedTimeUnixMs: processTimestamp * 1000,
-          },
-        },
-      }
-    })
-  },
-})
